move `server.ts` from `utils/` to `common/`
